requirements.txt
Flask
gunicorn
h5py
Jinja2
Keras
numpy
opencv-python
pandas
Pillow
request
scipy
tensorboard
tensorflow
tqdm
urllib3
utils